requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
pillow==11.0.0
numpy~=1.26.4
google-cloud-storage==2.14.0
google-cloud-firestore==2.14.0
python-multipart==0.0.9
requests==2.32.3
google-cloud-aiplatform==1.44.0
wsproto==1.2.0
opencv-python~=4.10.0.84
dlib~=19.24.6
mediapipe~=0.10.18